+++ src/main/scala/org/podval/tools/platform/Architecture.scala
@@ -1,0 +1,27 @@
+package org.podval.tools.platform
+
+enum Architecture derives CanEqual:
+  case i686
+  case x86_64
+  case amd64
+  case ppc64
+  case ppc64le
+  case s390x
+  case nacl
+  case aarch64
+  case armv6l
+  case armv7l
+  case armv8l
+
+object Architecture:
+  // Note: Gradle Node plugin's code claims that Java returns "arm" on all ARM variants.
+  def getNameFromEnvironment: String = System.getProperty("os.arch")
+
+  def getNameFromSystem: String = Exec.unameM
+
+  def getName: String = if Os.get.hasUname then getNameFromSystem else getNameFromEnvironment
+
+  def get: Architecture =
+    val name = getName
+    Architecture.values.find(_.toString.toLowerCase == name.toLowerCase)
+      .getOrElse(throw IllegalArgumentException(s"Unsupported architecture: $name"))