+++ src/main/scala/org/podval/tools/scalajs/ClosureCompilerSourceMapper.scala
@@ -2,8 +2,8 @@ package org.podval.tools.scalajs
 
 import com.google.debugging.sourcemap.SourceMapConsumerV3
 import com.google.debugging.sourcemap.proto.Mapping.OriginalMapping
-import org.opentorah.util.Files
 import org.podval.tools.testing.task.SourceMapper
+import org.podval.tools.util.Files
 import java.io.File
 
 final class ClosureCompilerSourceMapper(sourceMapFile: File) extends SourceMapper: