+++ src/main/scala/org/podval/tools/util/Strings.scala
@@ -1,0 +1,12 @@
+package org.podval.tools.util
+
+object Strings:
+  def split(what: String, on: Char): (String, Option[String]) = what.lastIndexOf(on) match
+    case -1 => (what, None)
+    case index => (what.substring(0, index), Some(what.substring(index+1)))
+
+  def drop(from: String, prefix: String): String =
+    if from.startsWith(prefix) then from.substring(prefix.length)
+    else throw IllegalArgumentException(s"String '$from' doesn't start with '$prefix'")
+
+  def prefix(prefix: String, what: Option[String]): String = what.fold("")(string => prefix + string)